Ticket #125 - Implementing tests for maximum number of pixels/channels/voxels. A few tweaks to remove compilation warnings.
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@1261 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/Detection/detection.hh
@@ -169,7 +169,7 @@ namespace duchamp
     };
 
     /// @brief How many channels does the Detection have? 
-    long   getNumChannels(){return Object3D::getNumDistinctZ();};
+    size_t   getNumChannels(){return Object3D::getNumDistinctZ();};
 
     /// @brief Is there at least the acceptable minimum number of channels in the Detection?  
     bool   hasEnoughChannels(int minNumber);
